various modernizations

+++ volume/dsp_volume_agmu.cpp
@@ -3,6 +3,8 @@
 #include "volume/db.h"
 #include "volume/dsp_helpers.h"
 
+#include "gsl/gsl_util"
+
 #include <algorithm>
 #include <limits>
 
@@ -19,7 +21,7 @@ void DspVolumeAGMU::process(gsl::span<int16_t> samples, int32_t channels)
     const auto old_peak = m_peak.load();
     const auto sample_count = samples.size();
     const auto frame_count = sample_count / channels;
-    auto peak = getPeak(samples.data(), sample_count);
+    auto peak = getPeak(samples.data(), gsl::narrow_cast<int32_t>(sample_count));
     peak = std::max(old_peak, peak);
     if (peak != old_peak)
     {
